[GUI] Convert screen-coords to GUI-coords

+++ src/client/gui/gui_constants.h
@@ -2,7 +2,7 @@
 
 #include <glm/vec2.hpp>
 
-constexpr unsigned GUI_WIDTH = 1920;
-constexpr unsigned GUI_HEIGHT = 1080;
+constexpr float GUI_WIDTH = 1920;
+constexpr float GUI_HEIGHT = 1080;
 
 const glm::vec2 GUI_VIEWPORT{GUI_WIDTH, GUI_HEIGHT};
